Added Support for bigger wordlists
When using the hasher program on bigger wordlists, it eats up the entire ram, so rather than processing a bigger file like over 1 GB, we rather split the files and then do process those subfiles and then merge them with sort while removing duplicates. All the user needs to do is chmod +x on this hashbigfiles.bash and run the script with appropriate arguments.